requirements.txt
asgiref==3.2.7
asn1crypto==1.3.0
astroid==2.3.3
boto3==1.12.39
botocore==1.15.39
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
conda==4.8.3
conda-package-handling==1.6.0
cryptography==2.8
Django==3.0.7
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
Flask==1.1.2
idna==2.8
importlib-metadata==1.6.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.5
pycosat==0.6.3
pycparser==2.19
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.22.0
ruamel-yaml==0.15.87
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
tqdm==4.42.1
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.1
wrapt==1.11.2
zipp==3.1.0
